split TabComponent into cpp & h - remove lambda injection from splitview->tabbar
