+++ Source/Tabbar.h
@@ -191,204 +191,48 @@ public:
     RecentlyOpenedListBox recentlyOpened;
 };
 
-class Canvas;
+class PluginEditor;
 class TabComponent : public TabbedComponent
     , public AsyncUpdater {
 
-    TextButton newButton = TextButton(Icons::Add);
+    TextButton newButton;
     WelcomePanel welcomePanel;
+    PluginEditor* editor;
 
 public:
-    std::function<void()> onTabMoved = []() {};
-    std::function<void()> onFocusGrab = []() {};
-    std::function<void(int)> onTabChange = [](int) {};
-    std::function<void()> newTab = []() {};
-    std::function<void()> openProject = []() {};
-    std::function<void(File&)> openProjectFile = [](File&) {};
-    std::function<void(int tabIndex, String const& tabName)> rightClick = [](int tabIndex, String const& tabName) {};
+    TabComponent(PluginEditor* editor);
 
-    TabComponent()
-        : TabbedComponent(TabbedButtonBar::TabsAtTop)
-    {
-        addAndMakeVisible(newButton);
-        newButton.getProperties().set("Style", "LargeIcon");
-        newButton.setColour(TextButton::buttonColourId, Colours::transparentBlack);
-        newButton.setTooltip("New patch");
-        newButton.onClick = [this]() {
-            newTab();
-        };
+    void onTabMoved();
+    void onFocusGrab();
+    void onTabChange(int tabIndex);
+    void newTab();
+    void openProject();
+    void openProjectFile(File& patchFile);
+    void rightClick(int tabIndex, String const& tabName);
 
-        addAndMakeVisible(welcomePanel);
+    void setSplitFocusIndex(int idx = -1);
 
-        welcomePanel.newButton.onClick = [this]() {
-            newTab();
-        };
+    void currentTabChanged(int newCurrentTabIndex, String const& newCurrentTabName) override;
+    void handleAsyncUpdate() override;
+    void resized() override;
 
-        welcomePanel.openButton.onClick = [this]() {
-            openProject();
-        };
+    void paint(Graphics& g) override;
 
-        welcomePanel.recentlyOpened.onPatchOpen = [this](File patchFile) {
-            openProjectFile(patchFile);
-        };
+    void paintOverChildren(Graphics& g) override;
 
-        setVisible(false);
-        setTabBarDepth(0);
-        tabs->addMouseListener(this, true);
-    }
+    void popupMenuClickOnTab(int tabIndex, String const& tabName) override;
 
-    void currentTabChanged(int newCurrentTabIndex, String const& newCurrentTabName) override
-    {
-        if (getNumTabs() == 0) {
-            setTabBarDepth(0);
-            getTabbedButtonBar().setVisible(false);
-            welcomePanel.show();
-        } else {
-            getTabbedButtonBar().setVisible(true);
-            welcomePanel.hide();
-            setTabBarDepth(30);
-        }
+    int getIndexOfCanvas(Canvas* cnv);
 
-        triggerAsyncUpdate();
-    }
+    Canvas* getCanvas(int idx);
 
-    void handleAsyncUpdate() override
-    {
-        onTabChange(getCurrentTabIndex());
-    }
+    Canvas* getCurrentCanvas();
 
-    void resized() override
-    {
-        int depth = getTabBarDepth();
-        auto content = getLocalBounds();
+    void mouseDown(MouseEvent const& e) override;
 
-        welcomePanel.setBounds(content);
-        newButton.setBounds(3, 0, depth, depth); // slighly offset to make it centred next to the tabs
+    void mouseDrag(MouseEvent const& e) override;
 
-        auto tabBounds = content.removeFromTop(depth).withTrimmedLeft(depth);
-        tabs->setBounds(tabBounds);
-
-        for (int c = 0; c < getNumTabs(); c++) {
-            if (auto* comp = getTabContentComponent(c)) {
-                if (auto* positioner = comp->getPositioner()) {
-                    positioner->applyNewBounds(content);
-                } else {
-                    comp->setBounds(content);
-                }
-            }
-        }
-    }
-
-    void paint(Graphics& g) override
-    {
-        g.setColour(findColour(PlugDataColour::tabBackgroundColourId));
-        g.fillRect(getLocalBounds().removeFromTop(30));
-    }
-
-    void paintOverChildren(Graphics& g) override
-    {
-        g.setColour(findColour(PlugDataColour::toolbarOutlineColourId));
-        g.drawLine(0, getTabBarDepth(), getWidth(), getTabBarDepth());
-
-        g.drawLine(0, 0, getWidth(), 0);
-        g.drawLine(0, 0, 0, getBottom());
-    }
-
-    void popupMenuClickOnTab(int tabIndex, String const& tabName) override
-    {
-        rightClick(tabIndex, tabName);
-    }
-
-    int getIndexOfCanvas(Canvas* cnv)
-    {
-        if (!cnv->viewport || !cnv->editor)
-            return -1;
-
-        for (int i = 0; i < getNumTabs(); i++) {
-            if (getTabContentComponent(i) == cnv->viewport) {
-                return i;
-            }
-        }
-
-        return -1;
-    }
-
-    Canvas* getCanvas(int idx)
-    {
-        auto* viewport = dynamic_cast<Viewport*>(getTabContentComponent(idx));
-
-        if (!viewport)
-            return nullptr;
-
-        return reinterpret_cast<Canvas*>(viewport->getViewedComponent());
-    }
-
-    Canvas* getCurrentCanvas()
-    {
-        auto* viewport = dynamic_cast<Viewport*>(getCurrentContentComponent());
-
-        if (!viewport)
-            return nullptr;
-
-        return reinterpret_cast<Canvas*>(viewport->getViewedComponent());
-    }
-
-    void mouseDown(MouseEvent const& e) override
-    {
-        tabWidth = tabs->getWidth() / std::max(1, getNumTabs());
-        clickedTabIndex = getCurrentTabIndex();
-        onFocusGrab();
-    }
-
-    void mouseDrag(MouseEvent const& e) override
-    {
-        // Don't respond to clicks on close button
-        if (dynamic_cast<TextButton*>(e.originalComponent))
-            return;
-        // Drag tabs to move their index
-        int const dragPosition = e.getEventRelativeTo(tabs.get()).x;
-        int const newTabIndex = (dragPosition < clickedTabIndex * tabWidth) ? clickedTabIndex - 1
-            : (dragPosition >= (clickedTabIndex + 1) * tabWidth)            ? clickedTabIndex + 1
-                                                                            : clickedTabIndex;
-        int const dragDistance = std::abs(e.getDistanceFromDragStartX());
-
-        if (dragDistance > 5) {
-            if ((tabs->contains(e.getEventRelativeTo(tabs.get()).getPosition()) || e.getDistanceFromDragStartY() < 0) && newTabIndex != clickedTabIndex && newTabIndex >= 0 && newTabIndex < getNumTabs()) {
-                moveTab(clickedTabIndex, newTabIndex, true);
-                clickedTabIndex = newTabIndex;
-                onTabMoved();
-                tabs->getTabButton(clickedTabIndex)->setVisible(false);
-            }
-
-            if (tabSnapshot.isNull() && (getParentWidth() != getWidth() || getNumTabs() > 1)) {
-                // Create ghost tab & hide dragged tab
-                auto* tabButton = tabs->getTabButton(clickedTabIndex);
-                currentTabBounds = tabButton->getBounds().translated(getTabBarDepth(), 0);
-
-                auto scale = Desktop::getInstance().getDisplays().getPrimaryDisplay()->scale;
-                tabSnapshot = Image(Image::PixelFormat::ARGB, tabButton->getWidth() * scale, tabButton->getHeight() * scale, true);
-
-                auto g = Graphics(tabSnapshot);
-                g.addTransform(AffineTransform::scale(scale));
-                getLookAndFeel().drawTabButton(*tabButton, g, false, false);
-
-                tabSnapshotBounds = currentTabBounds;
-                tabs->getTabButton(clickedTabIndex)->setVisible(false);
-            }
-            // Keep ghost tab within view
-            auto newPosition = Point<int>(std::clamp(currentTabBounds.getX() + getX() + e.getDistanceFromDragStartX(), 0, getParentWidth() - tabWidth), std::clamp(currentTabBounds.getY() + e.getDistanceFromDragStartY(), 0, getHeight() - tabs->getHeight()));
-            tabSnapshotBounds.setPosition(newPosition);
-            getParentComponent()->repaint();
-        }
-    }
-
-    void mouseUp(MouseEvent const& e) override
-    {
-        tabSnapshot = Image();
-        if (clickedTabIndex >= 0)
-            tabs->getTabButton(clickedTabIndex)->setVisible(true);
-        getParentComponent()->repaint(tabSnapshotBounds);
-    }
+    void mouseUp(MouseEvent const& e) override;
 
     Image tabSnapshot;
     Rectangle<int> tabSnapshotBounds;
